"ant playground" sometimes missing precajole.jar https://codereview.appspot.com/6930045
target PlaygroundBE copies precajole.jar into ant-war,
but precajole is not a prerequisite of PlaygroundBE,
so just running "ant playground" might not create precajole.jar
in time to be copied.

precajole is not a prereq of PlaygroundBE,
because PlaygroundBE is a prereq of AllTests,
and I don't want the slow precajole when doing "ant runtests".

so this fix is to copy precajole.jar in the playground rule,
since that's the rule that has precajole as a prereq.

R=kpreid2


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5191 360991c0-9f3b-0410-ba87-8b362a00edf4