make version-skew tests use static resources https://codereview.appspot.com/6921055
Currently, we have a number of tests that check various forms of
caja version mismatch, and these tests rely on starting the test
http server with different test-specific text substitution rules.
That makes it hard to manually examine or debug these tests.

So, this CL eliminates the dynamic rewrite. Instead, it statically
creates copies of the caja js files that have wrong version numbers,
and selects appropriate alternate files when running version-skew tests.

There are also a couple minor changes to make the browser tests run
a little more reliably with Chrome.

R=kpreid2


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5197 360991c0-9f3b-0410-ba87-8b362a00edf4