Explicitly prevent initEvent on already-dispatched events. https://codereview.appspot.com/7068050
It occurred to me to be concerned about initEvent being applied to
non-custom, browser-generated events, which I thought might allow
guests to modify events which then reach the host and confuse it. This
turns out to not be an actual risk because browsers have initEvent do
nothing if the event has already been dispatched. However, our
visible-only-to-the-guest emulation of initEvent for IE does not
implement this, so this is a virtualization improvement, on IE only.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5214 360991c0-9f3b-0410-ba87-8b362a00edf4