[APICHANGE] Stub out rather than rejecting non-simple dynamic string event handlers in ES5/3 mode. https://codereview.appspot.com/7309056
See <http://code.google.com/p/google-caja/issues/detail?id=1645> for
context. This does not resolve that issue but is a needed component.

Impact: In ES5/3 mode, setAttribute('onfoo', '<code other than a
function call>') used to be rejected and so any existing handler would
remain in effect. Now the existing handler will be replaced with a
nonfunctional handler.

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5280 360991c0-9f3b-0410-ba87-8b362a00edf4