squelch various firefox errors https://codereview.appspot.com/7226058
running browser tests on firefox generates a number of spurious
error messages in the console.

- there are a lot of messages about missing charset.  I didn't
see an easy way to make Jetty respond with the right charset,
so I just put <meta charset="utf-8"> in all the relevant files.

- es53-test-domado-dom-guest has a test that compares a uri
against 'localhost', which doesn't work with remote browsers.

- es53-test-unicode has problems in how it checks whether its
tests pass or not.

- cajajs-bare-test will append 'passed' to a script node,
causing a spurious eval error

R=kpreid2


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5312 360991c0-9f3b-0410-ba87-8b362a00edf4