fix race condition in webdriver shutdown https://codereview.appspot.com/7783043
The way we're using WebDriver often fails to exit cleanly because
of a race: we have a shutdown hook to close the browser instance,
and FirefoxDriver has a shutdown hook to delete the temporary
firefox profile it created. When we lose the race, firefox gets
confused by its missing profile dir, and sadness happens.

I spent way too long looking for other ways to do this. This is the
simplest I could think of, and it helps pave the way toward further
refactoring in the future.

R=kpreid2


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5320 360991c0-9f3b-0410-ba87-8b362a00edf4