Really fix whitelisting of cajaVM.es5ProblemReports. https://codereview.appspot.com/8064045
r5331 was slightly wrong (yes, this is a fix of a fix) in that it added
es5ProblemReports, let it be deleted by the whitelist, and then added it
again. This is largely harmless, but confusing and not warning-clean (it
causes the extra early copy to be reported as deleted).

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5332 360991c0-9f3b-0410-ba87-8b362a00edf4