Delete broken testScriptLoading <script> tests. https://codereview.appspot.com/8856044
Delete the tests of setting <script>.src in testScriptLoading in
es53-test-domita-special-guest.html and their supporting code. Also
rename testScriptLoading to its original name testCantLoadScript, since
it now only covers that.

As far as I can tell, these tests have almost never worked properly, are
now largely redundant, and have mostly been useful for confusing people.
Let me tell you their story, so no one has to rediscover it again:

The test of dynamically constructing a <script> and setting its
.src was introduced in r3406, when a feature was added where an
_unsandboxed_ script could be loaded iff the URI policy permitted it.
However, they used the unreliable assertAsynchronousRequirement
facility and so have never been effective at testing; this feature was
apparently lost at some point after this.

In r4506 as part of the transition to ES5/3 the test was moved, and
apparently misinterpreted, as a comment was rephrased to say 'cannot'
rather than 'can' load scripts, while the code was not changed. I do not
know whether the test was failing at this time.

In r4608 another assertion was added that a non-whitelisted script
cannot be so loaded. However, this was ineffective as the target
'[unproxied_]nonwhitelisted_script.js' did not exist to cause a detected
effect if it was erroneously loaded. Additionally, the async assertion
shared a global variable with the existing test so they would interfere.

In r4956, <script>.src was given meaning again, now loading _sandboxed_
scripts, and independent tests were added for this feature in
es53-test-external-script-guest.html.

R=felix8a@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5363 360991c0-9f3b-0410-ba87-8b362a00edf4