partial support for Safari testing https://codereview.appspot.com/9022043
With this CL, browser tests of safari work when connecting to a remote
webdriver.  Meaning:
  1. start "java -jar selenium-standalone*.jar"
  2. run "ant brtest -Dtest.browser=safari -Dtest.webdriver.url=http://127.0.0.1:4444/wd/hub"

the main fix in this CL is that window.open doesn't work in the
Safari webdriver.

There's also code to start a local SafariDriver, but that doesn't work
yet. (There's a weird protocol failure when trying to send a command
to the webdriver Safari extension, and I haven't figured out why yet.)

R=kpreid2



git-svn-id: http://google-caja.googlecode.com/svn/trunk@5384 360991c0-9f3b-0410-ba87-8b362a00edf4