Remove ExpandoProxyHandler. https://codereview.appspot.com/8700044
ExpandoProxyHandler gives no benefit in the current taming architecture;
it is a relic of times when we were supporting the notion of multiple
views of the same feral DOM node which should share properties, and this
capability is never used in the current architecture (and was only
present on proxy-supporting platforms).

Instead, the only Domado objects wrapped in proxies are those which need
it (e.g. <form> elements' nodelistiness).

The WeakMap support magic has a new home in ProxyHandler.

Rough benchmarks: in ES5 mode, jQuery test 'traversing' takes 87% of
its previous execution time, and 'manipulation' 80%.

R=ihab.awad@gmail.com



git-svn-id: http://google-caja.googlecode.com/svn/trunk@5393 360991c0-9f3b-0410-ba87-8b362a00edf4