Optimize proxy-based ArrayLike by implementing get trap. https://codereview.appspot.com/9456044
Implementing the get trap avoids constructing and discarding a
property descriptor whenever an element is retrieved from an
ArrayLike.

Results: 5% to 10% reduction in total execution time for jQuery and
jQuery UI tests on Firefox 20.0.

(Making note of preexisting test count discrepancies: 84 instead of 78
for jQuery UI 'position', 411 instead of 422 for jQuery 'attributes',
on Firefox 20.0. This change does not affect those counts.)

R=felix8a@gmail.com, erights@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5412 360991c0-9f3b-0410-ba87-8b362a00edf4