parse more css attribute match operators https://codereview.appspot.com/10236043
Right now the cajoler fails to parse stuff like
  p[x$=y] {}
and the parse failure causes it to throw out a lot more css than it should.

Fixing the error recovery is somewhat nontrivial. Adding the new operators
is trivial.

Note, this just allows the operators to parse, but rejects them as not
allowed, similar to the |= operator. So, the operators still don't work,
but we recover from that better.

R=ihab.awad



git-svn-id: http://google-caja.googlecode.com/svn/trunk@5445 360991c0-9f3b-0410-ba87-8b362a00edf4