downgrade css identifier error from FATAL to WARNING https://codereview.appspot.com/10347044
for some reason, a CSS selector like "input:-moz-placeholder" is 
a fatal error in es53, and it aborts processing.

I don't see a reason for this to be a fatal eror, so I'm downgrading
it to a warning. It's safe to do that because the warning still
marks that css rule as invalid, and it gets removed later.

this fixes https://code.google.com/p/google-caja/issues/detail?id=1768

R=ihab.awad



git-svn-id: http://google-caja.googlecode.com/svn/trunk@5452 360991c0-9f3b-0410-ba87-8b362a00edf4