trying to diagnose localserver port flakiness https://codereview.appspot.com/10761046
sometimes a browser test starts a local http server, then fails to
get the port number for that server, and the symptom is the
browser test tries to load http://localhost:-1/... which fails
since that isn't a valid URL.

the getPort failure is happening after the localserver is
started successfully (because I have a log message that the server
has started, and the timestamp is shortly before the message that
we're about to load the invalid URL.)

I don't have any good theories why this is happening. getting port=-1
is possible if the local server has stopped, but that isn't something
we're doing, there aren't any messages that look like the server is
stopping abnormally, and I can't think of any way for that to happen.

lacking any good theories, I'm adding some code that might help
diagnose the problem.

R=kpreid2



git-svn-id: http://google-caja.googlecode.com/svn/trunk@5469 360991c0-9f3b-0410-ba87-8b362a00edf4