workaround for selenium 2.33 vs firefox 23+ https://codereview.appspot.com/10963044
This fixes https://code.google.com/p/google-caja/issues/detail?id=1749

browser tests fail completely on firefox 23+ (beta and aurora)
because selenium's executeScript() fails.

I took a look at why executeScript is failing, and fixing it is
not particularly simple. So instead I'm reducing our dependency
on executeScript, because that's easy, and we really need to
test firefox beta (there's a new failure revealed by this fix).

We use executeScript in two ways:
1. querying firefox for build id, so we can tell different builds
of firefox apart. I've downgraded this to 'best effort'.

2. opening a new window for each browser test. There doesn't really
seem to be much value in doing that, and it complicates the test
driver in weird ways. I've stripped out much of that code, so we
now just re-use the same window instead of opening a new one.

R=kpreid2



git-svn-id: http://google-caja.googlecode.com/svn/trunk@5479 360991c0-9f3b-0410-ba87-8b362a00edf4