SES test and repair "FREEZING_BREAKS_WEAKMAP". https://codereview.appspot.com/12503044
V8 bug <https://code.google.com/p/v8/issues/detail?id=2829>:
If Object.prototype is in a WeakMap while Object.prototype is frozen
then WeakMap stops working properly (for objects inheriting from that
prototype). Add a test and repair which discards the broken WeakMap
in favor of our emulation.

(It is possible that there is a tweak to startSES's order of execution
which would leave the native WeakMap functional, but I don't want to try
to verify the correctness of such, at least unless we get some info on
the underlying cause from the V8 developers.)

Supporting changes (made to help find the above bug, but worthwhile
themselves even if not necessary for fixing it):
* hookupSES prints the error message as well as the object (which
  Chrome's console unfortunately doesn't do automatically).
* Guest content loaded as text/javascript did not receive JS globals
  in ES5 mode; fixed and tested.
* Domado is now constructed lazily when the first guest with DOM
  access is created (thus not at all if only JS is loaded).
* generic-host-page has an option to omit providing a div.

Fixes <https://code.google.com/p/google-caja/issues/detail?id=1831>.

R=erights@gmail.com, felix8a@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5535 360991c0-9f3b-0410-ba87-8b362a00edf4