add some missing css support https://codereview.appspot.com/12557045
This CL
- defines and allows box-sizing, overflow-wrap, word-wrap.
- fixes some of the background property defs.

There's a systematic problem with 'inherit' in our css-defs. A lot of
the css spec docs omit 'inherit' in describing individual properties
but have a statement like "All properties defined in this specification
also accept the inherit keyword as their value, but for readability it
has not been listed explicitly."

I've added a few of the missing inherits in this CL, but I haven't
yet tried to systematically add all the missing inherits. I suspect
it might be simpler to implicitly allow all properties to have the
value 'inherit', but I need to verify that's correct and/or safe.

R=mikesamuel



git-svn-id: http://google-caja.googlecode.com/svn/trunk@5537 360991c0-9f3b-0410-ba87-8b362a00edf4