generalize css vendor prefix for es53 https://codereview.appspot.com/12690043
This CL deletes the vendor-prefix css properties in the css whitelist.
To do that without breaking es53 users, this adds some vendor prefix
hxandling to the es53 css sanitizer.

This isn't quite feature parity with the es5 css sanitizer, because
es53 won't handle @-webkit-keyframes and other such cases, but that
isn't a new problem, and es53 is going away.

R=mikesamuel



git-svn-id: http://google-caja.googlecode.com/svn/trunk@5549 360991c0-9f3b-0410-ba87-8b362a00edf4