simplify css whitelist https://codereview.appspot.com/12899043
Our css rules try to make a distinction between pure css21 and
css-extensions, but in practice nobody wants to limit to css21, and
the separation makes it hard to audit the rules.

This CL eliminates the distinction between css21 and css-extensions,
so that there's only a single set of rules.

I have some more changes in mind, but this is a prerequisite,
so I'm trying to keep this change relatively simple. The diff is
unfortunately hard to audit, but it's mostly just combining the
css21 and css-extensions rules, re-alphabetizing, and removing the
duplicates (which are css21 rules superseded by css-extension rules)

So some of these rules are not quite what the css3 specs say.
Later, I'll fix up individual rules to conform to css3 reality.

R=mikesamuel



git-svn-id: http://google-caja.googlecode.com/svn/trunk@5565 360991c0-9f3b-0410-ba87-8b362a00edf4