Split scanner into two files, algorithm and configuration. https://codereview.appspot.com/12815049
test-scan-core.js contains the graph walking algorithm and is almost
independent of Caja. test-scan-guest.js contains all the data used to
scan the Caja environment.

There is one significant change beyond the split itself: the sort
ordering for property names is now a parameter to Scanner (and so can
live in test-scan-guest.js).

R=felix8a@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5576 360991c0-9f3b-0410-ba87-8b362a00edf4