Add early support for using a complete document as vdoc container. https://codereview.appspot.com/12775044
This support is documented as experimental, as we do not have complete
tests for functionality in the document case. Furthermore, we do not yet
use an un-virtualized <html> and <body>, which prevents some of the
desirable consequences of using a document.

* Modified caja.js to permit documents.
* Modified HtmlEmitter to permit documents and implement emitStatic
  without requiring the node to have an .innerHTML.
* Documented functionality in cajajs.xml.
* If in a document, emitCss puts stylesheets inside the guest tree,
  lacking anywhere else to put them. Note that in ES5 mode emitCss
  is being eliminated.
* Add event dispatch hooks to feral documents.

(The above last two changes I believe to eliminate the reasons why
caja.js previously had a restriction that the <div> container must be
in the same document as caja.js was loaded in.)

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5579 360991c0-9f3b-0410-ba87-8b362a00edf4