Document all 'canRepair: false' in repairES5. https://codereview.appspot.com/13242047
Per <https://code.google.com/p/google-caja/issues/detail?id=1836>:
In SES, some problems are marked canRepair: false, not because no
possible repair exists, but because we no longer expect to see that
problem and so the repair code is not worth keeping around (r5524).

Add a comment explaining in each case why no repair is present, as a
best guess based on the history of each problem record.

R=erights@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5580 360991c0-9f3b-0410-ba87-8b362a00edf4