Validate severity name passed to ses.ok. https://codereview.appspot.com/13694048
Previously, ses.ok('not a valid name') was treated as equivalent to
ses.ok(), which is dangerous if, for example, the maxAcceptableSeverity
is set high and the intent of the ses.ok(...) call was to make a choice
based on meeting a lower severity (as caja.js does). Now, passing an
unknown severity name will throw an error.

Fixes <https://code.google.com/p/google-caja/issues/detail?id=1758>.

Supporting changes:
* browser-test-case.html no longer requires an es5= parameter, but will
  fail if anything which would depend on it is called while it is
  unspecified. This allows it to be used for tests that do not use
  caja.js at all, such as this new test for SES.
* Tweaked ses-tests.json catalog to use inheritance/grouping.
* Note: severityNameToLevel was misnamed and entirely unused.

R=erights@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5603 360991c0-9f3b-0410-ba87-8b362a00edf4