deprecate com.google.caja.util.Maps https://codereview.appspot.com/13861043
com.google.common.collect.Maps from Guava does the same thing.

The comment at the top of the Maps class is referring to a very
old version of the google collections library, which was fixed
to be like this Maps class, and then absorbed into Guava.

the deletion of one of the newTreeBag factory methods is because
the necessary generics turn out to be weird, because of bugs in
various versions of javac. since we're not using that method
I just deleted it instead.

R=kpreid2



git-svn-id: http://google-caja.googlecode.com/svn/trunk@5604 360991c0-9f3b-0410-ba87-8b362a00edf4