Remove special handling of "simple" event handler attributes. https://codereview.appspot.com/22060043
Since eval is now always available, the notion of simple event handlers
which are rewritten and executed without eval is unnecessary complexity,
which we now remove.

Note this has an additional cost: event handler attributes are compiled
and kept around in the domicile.handlers table and never deleted from
it, whereas simple event handlers would be stored inline in the
rewritten handler and thus not accumulate garbage if repeatedly created.
This could be avoided by storing the code inline as well (as a string)
and either recompiling each time the handler is invoked or using a
bounded cache.

This change affects only event handlers expressed as attributes
(<span onclick="..."> or el.setAttribute('onclick', ...)), not those
expressed as functions (el.onclick = function () {...}).

R=felix8a@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5623 360991c0-9f3b-0410-ba87-8b362a00edf4