Make scanner self-test omit prototypes. https://codereview.appspot.com/24710043
test-scan-guest testScanner is unnecessarily fragile because it exposes
more ordinary object structure than just what was tested for, possibly
reacting to changes in the scan or problem reporting than what it was
looking for. Now the test object graph is run through detach(), which
reproduces the graph with no [[Prototype]]s to keep it to the explicitly
written structure.

R=felix8a@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5628 360991c0-9f3b-0410-ba87-8b362a00edf4